Integration with backend. Backend Improvements.
Frontend: at least was added. Consists of 3 pages.
Landing, Login/Register, Chat
All seems to be working with backend, because there are successful requests on backend server

Backend: works.
Added a system prompt for LLM, this tweaks LLM's responses (and I hope we get what we need)
Added a bit of asynchronous stuff (because we couldn't use anything when LLM is generating something)

Co-Authored-By: STIDchannel <52779739+STIDchannel@users.noreply.github.com>
Co-Authored-By: Anastasia Zabolotnay <197951069+Anastasia11111111114@users.noreply.github.com>
Co-Authored-By: ProstoAndriuha <209294428+ProstoAndriuha@users.noreply.github.com>